pyproject.toml
[tool.poetry]
name = "lidar-optimizer"
version = "0.1.0"
description = ""
authors = ["Surikkat <Surikkatik@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
numpy = "^1.26.4"
natsort = "^8.4.0"
opencv-python = "^4.9.0.80"
pillow = "^10.2.0"
pandas = "^2.2.1"
tensorflow = "^2.15.0.post1"
scikit-learn = "^1.4.1.post1"
segmentation-models = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
